refactor(test): separate test folders by app
- 기능별 테스트의 위치를 분할했습니다
